requirements/base.txt
Django==3.0.1
django-cors-headers==3.2.0
django-environ==0.4.5
djangorestframework==3.11.0
psycopg2-binary==2.8.4
uwsgi==2.0.18
django-admin-sortable==2.2
django-crispy-forms==1.8.1
django-allauth==0.41.0
-e git://github.com/jay11377/django-suit@v2#egg=django-suit
django-autoslug==1.9.6
